+++ src/Plainion.Windows/Controls/Text/NoteBook.xaml.cs
@@ -1,5 +1,8 @@
-﻿using System.Windows;
+﻿using System.Diagnostics;
+using System.Linq;
+using System.Windows;
 using System.Windows.Controls;
+using System.Collections.ObjectModel;
 
 namespace Plainion.Windows.Controls.Text
 {
@@ -11,12 +14,41 @@ namespace Plainion.Windows.Controls.Text
         }
 
         public static readonly DependencyProperty DocumentStoreProperty = DependencyProperty.Register("DocumentStore",
-            typeof(IDocumentStore), typeof(NotePad), new PropertyMetadata(null, OnDocumentStoreChanged));
+            typeof(IDocumentStore), typeof(NoteBook), new PropertyMetadata(null, OnDocumentStoreChanged));
 
         private static void OnDocumentStoreChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
             var self = (NoteBook)d;
-            //self.myNavigation.Tree = BuildTree(self.DocumentStore);
+
+            var root = self.myNavigation.Root;
+
+            root.Children.Clear();
+
+            foreach(var process in Process.GetProcesses())
+            {
+                var processNode = new NavigationNode
+                {
+                    Parent = root,
+                    Id = process.Id.ToString(),
+                    Name = process.ProcessName,
+                    IsDragAllowed = false
+                };
+                root.Children.Add(processNode);
+
+                var children = process.Threads
+                    .OfType<ProcessThread>()
+                    .Select(t => new NavigationNode
+                    {
+                        Parent = processNode,
+                        Id = t.Id.ToString(),
+                        Name = "unknown",
+                        IsDropAllowed = false
+                    });
+                foreach(var child in children)
+                {
+                    processNode.Children.Add(child);
+                }
+            }
         }
 
         public IDocumentStore DocumentStore