+++ ChangeLog.md
@@ -1,7 +1,6 @@
 ## 1.23.0 - unreleased
 
-- TreeEditor
-  - context menu handling improved
+- Mvvm.DelegateCommand, Mvvm.BindableBase added to avoid referencing Prism library for small projects
 
 ## 1.22.0 - 2016-09-06